aktualisiert

+++ bin/mailrun.sh
@@ -6,16 +6,15 @@
 # - every 2 minutes a partial synch of the INBOX'es is done, if AC-power is
 #   connected
 # - runs a full synch after standby
+# - trys to synch, only when network connected
 # - the log is saved to /tmp/mailrun-log
 # - the routine _monitor() checks, if offlineimap is alive, by checking wether
 #   the log file changes
 #
-# Last modified: So Jul 21, 2013  11:05
+# Last modified: Di Jul 23, 2013  12:02
 
 PID=$$
-
 PIDFILE=/tmp/mailrun-sh-pid
-TIMEFILE=/tmp/mailrun-sh-lastrun
 LOGFILE=/tmp/mailrun-sh-log
 
 pid=$(pgrep -f "/usr/bin/offlineimap")
@@ -25,11 +24,10 @@ if [[ ${pid} -gt 0 ]] ; then
 fi
 if [ -f ${PIDFILE} ] && [ -d "/proc/`cat $PIDFILE`" ]; then
   echo "mailrun.sh is already running"
-  exit 2 # or kill the running instance?
+  exit 2 # TODO: or kill the running instance?
 fi
 
-rm -f $TIMEFILE
-echo $PID > $PIDFILE
+echo $PID >$PIDFILE
 echo "**** started at $(date) ****" >$LOGFILE 2>&1
 
 # Monitor offlineimap #######################################################
@@ -41,28 +39,35 @@ _monitor() {
     if [ -f $LOGFILE ]; then
       if test `find ${LOGFILE} -mmin +5`; then
         # dead for more then 5 Minutes
-        pid=$(pgrep -f "/usr/bin/offlineimap")
-        if [[ ${pid} -gt 0 ]] ; then
-          echo "#############################################"
-          echo "#############################################"
-          echo "Offlineimap has hung with pid ${pid} (killed)"
-          pkill -9 -f 'offlineimap\>.*-o'
-          echo "#############################################"
-          echo "#############################################"
+        if [[ $(pgrep -f "/usr/bin/offlineimap") -gt 0 ]] ; then
+          sleep 20 # just wait some more, bevore killing offlineimap
+          if test `find ${LOGFILE} -mmin +5`; then 
+            # test, if it is realy dead
+            pid=$(pgrep -f "/usr/bin/offlineimap")
+            if [[ ${pid} -gt 0 ]] ; then
+              # kill offlineimap
+              pkill -9 -f 'offlineimap\>.*-o'
+              echo "#############################################"
+              echo "#############################################"
+              echo "Offlineimap has hung with pid ${pid} (killed)"
+              echo "#############################################"
+              echo "#############################################"
+            fi
+          fi
         fi
       fi
     fi
   done >>$LOGFILE 2>&1
 }
 _monitor &
-#MONITOR_PID=$!
 trap 'kill $(jobs -p)' EXIT
 
+LASTRUN=0
 sleeptime=15 # sleep only 15s on the first run
 # main loop #################################################################
 while true; do
   sleep $sleeptime
-  sleeptime=120
+  sleeptime=$((2*60))
   # check, if connected #####################################################
   if ! ping -c1 cloud.github.com > /dev/null 2>&1; then
     echo "**** not connected (ping) ****" >&2
@@ -80,8 +85,8 @@ while true; do
     echo "#############################################"
   else
     CURR=$(date '+%s')
-    if [ ! -f $TIMEFILE ] || [ $(($CURR-$(cat $TIMEFILE))) -gt $((600)) ]; then
-      echo "$CURR" >$TIMEFILE
+    if [ $(($CURR-$LASTRUN)) -gt $((10*60)) ]; then
+      LASTRUN=$CURR
       /usr/bin/offlineimap -o -u ttyui
     else
       if [[ $(acpi -a | grep -c on-line) == "1" ]]; then
@@ -89,10 +94,10 @@ while true; do
       else
         echo "**** no shortsynch, because no AC ****"
       fi
+      echo "**** $((10-($CURR-$LASTRUN)/60))m till next full run ****"
     fi
     echo "**** Offlineimap is ready (at $(date)) ****"
   fi
 done >>$LOGFILE 2>&1
 
 rm $PIDFILE
-rm $TIMEFILE
